convert remaining paster commands to click
